+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Changed `join_nulls` option to `nulls_equal` for `join` method
 - Changed `strict` option to `check_dtypes` for `equals` method
 - Changed `quote` option to `quote_char` for `write_csv` method
+- Changed `data_pagesize_limit` option to `data_page_size` for `sink_parquet` method
 - Changed `unit` option to `time_unit` for `from_epoch` method
 - Removed `warn_if_unsorted` option from `rolling_*_by` methods
 - Removed `in_place` option from `rename` method